refactor(process): 拆出 ProcessPage 表格、详情抽屉与交互 hook
